+++ dlls/d3dx10_43/tests/d3dx10.c
@@ -25,6 +25,14 @@
 #include <stdint.h>
 #include <assert.h>
 
+static const D3DXMATRIX identity =
+{
+    ._11 = 1.0f,
+    ._22 = 1.0f,
+    ._33 = 1.0f,
+    ._44 = 1.0f,
+};
+
 #define SWAPCHAIN_FLAG_SHADER_INPUT             0x1
 
 struct swapchain_desc
@@ -5047,21 +5055,19 @@ static void test_dxt10_dds_header_image_info(void)
                 D3D10_RESOURCE_DIMENSION_TEXTURE2D
             }
         },
-        /* Resource dimension is validated for cube textures. */
-        {
-            0, 4, 4, 1, (4 * 4), 1, 0, 0,
-            { DXGI_FORMAT_R8G8B8A8_UNORM, D3D10_RESOURCE_DIMENSION_TEXTURE3D, DDS_RESOURCE_MISC_TEXTURECUBE, 2, 0, },
-            (4 * 4 * 4 * 12), { E_FAIL }
-        },
-        /*
-         * 10.
-         * 1D Texture cube, invalid.
-         */
-        {
-            0, 4, 4, 1, (4 * 4), 1, 0, 0,
-            { DXGI_FORMAT_R8G8B8A8_UNORM, D3D10_RESOURCE_DIMENSION_TEXTURE1D, DDS_RESOURCE_MISC_TEXTURECUBE, 2, 0, },
-            (4 * 4 * 4 * 12), { E_FAIL }
-        },
+        /* Resource dimension is not necessarily validated for cube textures,
+         * it sporadically crashes on Windows. */
+        /* { */
+        /*     0, 4, 4, 1, (4 * 4), 1, 0, 0, */
+        /*     { DXGI_FORMAT_R8G8B8A8_UNORM, D3D10_RESOURCE_DIMENSION_TEXTURE3D, DDS_RESOURCE_MISC_TEXTURECUBE, 2, 0, }, */
+        /*     (4 * 4 * 4 * 12), { E_FAIL } */
+        /* }, */
+        /* 1D Texture cube, invalid. */
+        /* { */
+        /*     0, 4, 4, 1, (4 * 4), 1, 0, 0, */
+        /*     { DXGI_FORMAT_R8G8B8A8_UNORM, D3D10_RESOURCE_DIMENSION_TEXTURE1D, DDS_RESOURCE_MISC_TEXTURECUBE, 2, 0, }, */
+        /*     (4 * 4 * 4 * 12), { E_FAIL } */
+        /* }, */
     };
     D3DX10_IMAGE_INFO info;
     unsigned int i;
@@ -6155,41 +6161,33 @@ static void test_dxt_formats(void)
 
 static void test_srgb_filter_flags(void)
 {
+    /*
+     * An input value > 1.0f being converted from SRGB to linear produces
+     * quite a few different values depending on the SDK version, and
+     * sometimes changes run to run. Presumably it's reading beyond the end of
+     * a LUT, values above ~106.0f will cause a crash on all versions and
+     * bitnesses.
+     */
     static const float test_float4_srgb_in[] =
     {
         0.09f, 0.1f,  0.2f, 1.0f,
         0.30f, 0.4f,  0.5f, 2.0f,
         0.60f, 0.7f,  0.8f, 3.0f,
-        0.90f, 1.5f, -1.0f, 4.0f,
+        0.90f, 1.0f, -1.0f, 4.0f,
     };
     static const float test_float4_srgb_in_expected[] =
     {
         5.00732847e-003,  6.27983455e-003, 2.89932229e-002, 1.00000000e+000,
         7.07391128e-002,  1.33206353e-001, 2.17635408e-001, 2.00000000e+000,
         3.25037479e-001,  4.56263810e-001, 6.12064898e-001, 3.00000000e+000,
-        7.93109715e-001, -2.24207754e-044, 1.00000000e+000, 4.00000000e+000,
+        7.93109715e-001,  1.00000000e+000, 1.00000000e+000, 4.00000000e+000,
     };
     static const float test_float4_srgb_in_expected_32[] =
     {
         5.00732893e-003, 6.27983361e-003, 2.89932191e-002, 1.00000000e+000,
         7.07391202e-002, 1.33206338e-001, 2.17635408e-001, 2.00000000e+000,
         3.25037509e-001, 4.56263840e-001, 6.12064838e-001, 3.00000000e+000,
-        /*
-         * On 32-bit d3dx10+, an input value of 1.5f being converted from SRGB
-         * to linear produces quite a few different values depending on the
-         * SDK version. Presumably it's reading beyond the end of a LUT,
-         * values above ~106.0f will cause a crash on all versions and
-         * bitnesses.
-         */
-#if D3DX10_SDK_VERSION < 35
-        7.93109715e-001, 3.48807693e-001, 1.00000000e+000, 4.00000000e+000
-#elif D3DX10_SDK_VERSION < 37
-        7.93109715e-001, 3.56686294e-001, 1.00000000e+000, 4.00000000e+000
-#elif D3DX10_SDK_VERSION < 40
-        7.93109715e-001, 3.64580810e-001, 1.00000000e+000, 4.00000000e+000
-#else
-        7.93109715e-001, 3.33099246e-001, 1.00000000e+000, 4.00000000e+000
-#endif
+        7.93109715e-001, 1.00000000e+000, 1.00000000e+000, 4.00000000e+000
     };
     static const float test_float4_srgb_out[] =
     {
@@ -7835,13 +7833,6 @@ static void test_sprite(void)
     D3DXMATRIX mat, mat2;
     ULONG refcount;
     HRESULT hr;
-    static const D3DXMATRIX identity =
-    {
-        ._11 = 1.0f,
-        ._22 = 1.0f,
-        ._33 = 1.0f,
-        ._44 = 1.0f,
-    };
 
     if (!(device = create_device()))
     {
@@ -7916,12 +7907,18 @@ static void test_sprite(void)
     ok(!memcmp(&mat, &mat2, sizeof(mat)), "Unexpected matrix.\n");
 
     /* View transform */
+    hr = ID3DX10Sprite_GetViewTransform(sprite, NULL);
+    ok(hr == E_FAIL, "Unexpected hr %#lx.\n", hr);
+
+    memset(&mat, 0, sizeof(mat));
+    hr = ID3DX10Sprite_GetViewTransform(sprite, &mat);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+    ok(!memcmp(&mat, &identity, sizeof(mat)), "Unexpected view transform.\n");
+
     hr = ID3DX10Sprite_SetViewTransform(sprite, NULL);
-    todo_wine
     ok(hr == E_FAIL, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_SetViewTransform(sprite, &mat);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     /* Begin */
@@ -7933,7 +7930,6 @@ static void test_sprite(void)
 
     /* Flush/End */
     hr = ID3DX10Sprite_Flush(sprite);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_End(sprite);
@@ -7968,11 +7964,9 @@ static void test_sprite(void)
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_Flush(sprite);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_Flush(sprite);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_End(sprite);
@@ -7991,7 +7985,6 @@ static void test_sprite(void)
     ok(get_refcount(srv1) > refcount, "Unexpected refcount.\n");
 
     hr = ID3DX10Sprite_Flush(sprite);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
     ok(get_refcount(srv1) == refcount, "Unexpected refcount.\n");
 
@@ -8172,27 +8165,113 @@ static void test_sprite_render(void)
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     hr = ID3DX10Sprite_Flush(sprite);
-    todo_wine
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
     hr = ID3DX10Sprite_End(sprite);
     ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
 
     color = get_texture_color(test_context.backbuffer, 160, 120);
-    todo_wine
     ok(compare_color(color, 0xff0000ff, 0), "Got unexpected color 0x%08x.\n", color);
     color = get_texture_color(test_context.backbuffer, 480, 120);
-    todo_wine
     ok(compare_color(color, 0xffff00ff, 0), "Got unexpected color 0x%08x.\n", color);
     color = get_texture_color(test_context.backbuffer, 160, 360);
-    todo_wine
     ok(compare_color(color, 0xffff0000, 0), "Got unexpected color 0x%08x.\n", color);
     color = get_texture_color(test_context.backbuffer, 480, 360);
-    todo_wine
     ok(compare_color(color, 0xff00ffff, 0), "Got unexpected color 0x%08x.\n", color);
 
-    ID3DX10Sprite_Release(sprite);
+    /* Immediate draws with sprites in the batch */
+    ID3D10Device_ClearRenderTargetView(device, test_context.backbuffer_rtv, clear);
+
+    hr = ID3DX10Sprite_Begin(sprite, 0);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+
+    color = get_texture_color(test_context.backbuffer, 160, 120);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 120);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 160, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+
+    D3DXMatrixTranslation(&sprite_desc.matWorld, -0.5f, 0.5f, 0.0f);
+    sprite_desc.TexCoord.x = 0.0f;
+    sprite_desc.TexCoord.y = 0.0f;
+    sprite_desc.TexSize.x = 0.25f;
+    sprite_desc.TexSize.y = 0.25f;
+    hr = ID3DX10Sprite_DrawSpritesBuffered(sprite, &sprite_desc, 1);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+
+    color = get_texture_color(test_context.backbuffer, 160, 120);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 120);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 160, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+
+    D3DXMatrixTranslation(&sprite_desc.matWorld, 0.5f, 0.5f, 0.0f);
+    sprite_desc.TexCoord.x = 0.7f;
+    sprite_desc.TexCoord.y = 0.0f;
+    sprite_desc.TexSize.x = 0.25f;
+    sprite_desc.TexSize.y = 0.25f;
+    hr = ID3DX10Sprite_DrawSpritesImmediate(sprite, &sprite_desc, 1, sizeof(sprite_desc), 0);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+
+    color = get_texture_color(test_context.backbuffer, 160, 120);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 120);
+    ok(compare_color(color, 0xffff00ff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 160, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 480, 360);
+    ok(compare_color(color, 0xffffffff, 0), "Got unexpected color 0x%08x.\n", color);
+
+    hr = ID3DX10Sprite_End(sprite);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+
     ID3D10Texture2D_Release(texture);
     ID3D10ShaderResourceView_Release(srv);
+
+    /* Testing sampler filtering mode */
+    texture_desc.Width = 16;
+    texture_desc.Height = 16;
+
+    resource_data.pSysMem = a8r8g8b8_16_16;
+    resource_data.SysMemPitch = texture_desc.Width * 4;
+    resource_data.SysMemSlicePitch = 0;
+    hr = ID3D10Device_CreateTexture2D(device, &texture_desc, &resource_data, &texture);
+    ok(hr == S_OK, "Failed to create texture, hr %#lx.\n", hr);
+
+    hr = ID3D10Device_CreateShaderResourceView(device, (ID3D10Resource *)texture, NULL, &srv);
+    ok(hr == S_OK, "Failed to create srv, hr %#lx.\n", hr);
+
+    ID3D10Device_ClearRenderTargetView(device, test_context.backbuffer_rtv, clear);
+
+    sprite_desc.TexCoord.x = 0.0f;
+    sprite_desc.TexCoord.y = 0.0f;
+    sprite_desc.TexSize.x = 1.0f;
+    sprite_desc.TexSize.y = 1.0f;
+    sprite_desc.pTexture = srv;
+    sprite_desc.matWorld = identity;
+
+    hr = ID3DX10Sprite_Begin(sprite, 0);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+    hr = ID3DX10Sprite_DrawSpritesImmediate(sprite, &sprite_desc, 1, 0, 0);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+    hr = ID3DX10Sprite_End(sprite);
+    ok(hr == S_OK, "Unexpected hr %#lx.\n", hr);
+
+    color = get_texture_color(test_context.backbuffer, 320, 235);
+    ok(compare_color(color, 0x7b7b7b7b, 1), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 320, 240);
+    ok(compare_color(color, 0x80808080, 1), "Got unexpected color 0x%08x.\n", color);
+    color = get_texture_color(test_context.backbuffer, 320, 245);
+    ok(compare_color(color, 0x85858585, 1), "Got unexpected color 0x%08x.\n", color);
+
+    ID3D10Texture2D_Release(texture);
+    ID3D10ShaderResourceView_Release(srv);
+    ID3DX10Sprite_Release(sprite);
     release_test_context(&test_context);
 }